+++ packages/terra-menu/src/Menu.jsx
@@ -1,15 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import Popup from 'terra-popup';
-import SlideGroup from 'terra-slide-group';
 import classNames from 'classnames/bind';
 import 'terra-base/lib/baseStyles';
 import MenuItem from './MenuItem';
 import MenuItemGroup from './MenuItemGroup';
 import MenuDivider from './MenuDivider';
 import MenuContent from './_MenuContent';
-import MenuWidths from './_MenuWidths';
-import MenuUtils from './_MenuUtils';
 import styles from './Menu.scss';
 
 const cx = classNames.bind(styles);
@@ -49,9 +46,9 @@ const propTypes = {
   isOpen: PropTypes.bool,
   /**
    * A string representation of the width in px, limited to:
-   * 160, 240, 320, 640, 960, 1280, 1760
+   * 160, 240, 320, 640, 960, 1280, 1760 or auto
    */
-  contentWidth: PropTypes.oneOf(['160', '240', '320', '640', '960', '1280', '1760']),
+  contentWidth: PropTypes.oneOf(Object.keys(Popup.Opts.widths)),
   /**
    * Indicates if the menu should have an center aligned arrow displayed on dropdown.
    * Otherwise, the menu will display without an arrow and right aligned.
@@ -66,20 +63,14 @@ const defaultProps = {
   contentWidth: '240',
 };
 
-const MENU_PADDING_TOP = 6;
-const MENU_PADDING_BOTTOM = 6;
-const MENU_ITEM_HEIGHT = 23;
-const MENU_DIVIDER_HEIGHT = 10;
-
 class Menu extends React.Component {
   constructor(props) {
     super(props);
-    this.getContentHeight = this.getContentHeight.bind(this);
+    this.setPageDimensions = this.setPageDimensions.bind(this);
     this.push = this.push.bind(this);
     this.pop = this.pop.bind(this);
     this.state = { stack: [this] };
   }
-
 
   componentWillReceiveProps(nextProps) {
     if ((this.props.isOpen && !nextProps.isOpen) || this.props.children.length !== nextProps.children.length) {
@@ -87,21 +78,16 @@ class Menu extends React.Component {
     }
   }
 
-  getContentHeight() {
-    let itemCount = 0;
-    let dividerCount = 0;
-
-    React.Children.forEach(this.props.children, (child) => {
-      if (child.props.children && child.props.children.length > 0) {
-        itemCount += child.props.children.length;
-      } else if (child.type === <MenuDivider />.type) {
-        dividerCount += 1;
-      } else {
-        itemCount += 1;
+  setPageDimensions(node) {
+    if (node) {
+      this.pageHeight = node.clientHeight;
+      if (this.props.contentWidth === 'auto') {
+        this.pageWidth = node.clientWidth;
       }
-    });
-
-    return (itemCount * MENU_ITEM_HEIGHT) + (dividerCount * MENU_DIVIDER_HEIGHT) + MENU_PADDING_TOP + MENU_PADDING_BOTTOM;
+    } else {
+      this.pageHeight = undefined;
+      this.pageWidth = undefined;
+    }
   }
 
   pop() {
@@ -130,25 +116,13 @@ class Menu extends React.Component {
       contentWidth,
       ...customProps
     } = this.props;
-    const contentHeight = this.getContentHeight();
-    const popupHeight = MenuUtils.getPopupHeight(contentHeight);
-    const boundingFrame = this.props.boundingRef ? this.props.boundingRef() : undefined;
-    const isFullScreen = MenuUtils.isFullScreen(boundingFrame, popupHeight, MenuWidths[this.props.contentWidth]);
-
-    const isSubMenu = this.state.stack.length > 1;
-    const contentClass = cx([
-      { submenu: isSubMenu },
-      { 'main-menu': !isSubMenu },
-      { fullscreen: isFullScreen },
-      classNameContent,
-    ]);
-
     const arrowClass = cx([
       'arrow',
-      { submenu: isSubMenu },
+      { submenu: this.state.stack.length > 1 },
       classNameArrow,
     ]);
 
+    const visiblePage = this.state.stack.length - 1;
     const slides = this.state.stack.map((item, index) => (
       <MenuContent
         // eslint-disable-next-line react/no-array-index-key
@@ -157,7 +131,14 @@ class Menu extends React.Component {
         onRequestNext={this.push}
         onRequestBack={this.pop}
         onRequestClose={this.props.onRequestClose}
+        isHidden={index !== visiblePage}
+        fixedHeight={this.pageHeight}
+        fixedWidth={this.pageWidth}
+        contentWidth={Popup.Opts.widths[contentWidth]}
+        refCallback={visiblePage === 0 ? this.setPageDimensions : null}
         index={index}
+        boundingRef={boundingRef}
+        isFocused={index === visiblePage}
       >
         {item.props.children || item.props.subMenuItems}
       </MenuContent>
@@ -168,18 +149,20 @@ class Menu extends React.Component {
         {...customProps}
         boundingRef={boundingRef}
         isArrowDisplayed={isArrowDisplayed}
+        attachmentBehavior="flip"
         contentAttachment={isArrowDisplayed ? 'top center' : 'top right'}
-        contentHeight={popupHeight.toString()}
+        contentHeight="auto"
         contentWidth={this.props.contentWidth}
         classNameArrow={arrowClass}
-        classNameContent={contentClass}
+        classNameContent={classNameContent}
         classNameOverlay={classNameOverlay}
         isOpen={isOpen}
         onRequestClose={onRequestClose}
         targetRef={targetRef}
         isHeaderDisabled
+        isContentFocusDisabled
       >
-        <SlideGroup items={slides} />
+        {slides}
       </Popup>
     );
   }
@@ -190,5 +173,8 @@ Menu.defaultProps = defaultProps;
 Menu.Item = MenuItem;
 Menu.ItemGroup = MenuItemGroup;
 Menu.Divider = MenuDivider;
+Menu.Opts = {
+  widths: Popup.Opts.widths,
+};
 
 export default Menu;